Fixed Value Conversion Warnings (double -> float)

+++ Observer/Maps/MapSelectViewController.m
@@ -223,7 +223,7 @@
         cell.percentComplete = map.downloadPercentComplete;
         map.downloadProgressAction = ^(double bytesWritten, double bytesExpected) {
             dispatch_async(dispatch_get_main_queue(), ^{
-                cell.percentComplete = bytesWritten/bytesExpected;
+                cell.percentComplete = (float)(bytesWritten/bytesExpected);
             });
         };
         return cell;
